bug：测试添加日志功能时，不停的向数据库里面添加记录， tomorrow, it will be fixed

+++ drp/src/main/java/com/lyu/drp/sysmanage/entity/User.java
@@ -16,7 +16,7 @@ public class User implements Serializable {
 	
 	private Long userId;
 
-    private Integer deptId;
+    private Long deptId;
 
     private String loginName;
 
@@ -48,11 +48,11 @@ public class User implements Serializable {
 		this.userId = userId;
 	}
 
-	public Integer getDeptId() {
+	public Long getDeptId() {
 		return deptId;
 	}
 
-	public void setDeptId(Integer deptId) {
+	public void setDeptId(Long deptId) {
 		this.deptId = deptId;
 	}
